Bug fixes and updates

+++ src/main/java/io/github/timpcunningham/anax/commands/world/delete/DeleteCommand.java
@@ -9,6 +9,7 @@ import io.github.timpcunningham.anax.utils.chat.Lang;
 import io.github.timpcunningham.anax.utils.command.CommandQueue;
 import io.github.timpcunningham.anax.utils.player.CommandUtils;
 import io.github.timpcunningham.anax.utils.world.WorldUtils;
+import io.github.timpcunningham.anax.world.AnaxWorldManagement;
 import io.github.timpcunningham.anax.world.tables.AnaxWorld;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
@@ -25,6 +26,10 @@ public class DeleteCommand {
         AnaxWorld world = CommandUtils.validateWorldLoaded(sender, player.getWorld().getName());
 
         WorldUtils.assertCanManage(player, world);
+
+        if(AnaxWorldManagement.getInstance().isDefaultWorld(world)) {
+            throw new LocalizedCommandException(player, Lang.WORLD_DEFAULT_DENY);
+        }
 
         Chat.alertPlayer(player, Lang.COMMAND_CONFIRM, Lang.FORMAT_GLOBAL_ALERT, 15);
         CommandQueue.getInstance().add(player.getUniqueId(), 15, new DeleteCallback(), world, player);
